application_architecture_distribuee/StreamingServer/app/requirements.txt
Flask
zeroc-ice
pymongo
python-dotenv
dnspython